Added StateMachine API. Refactored to use the StateMachine internally

+++ examples/src/test/java/example/inheritance/ConcreteOddTest.java
@@ -37,13 +37,13 @@ public class ConcreteOddTest extends OddBaseTest {
     @TestEngine.Test
     public void test1() {
         System.out.println("test1(" + integerArgument + ")");
-        assertThat((Integer) integerArgument.value() % 2).isOdd();
+        assertThat(integerArgument.value() % 2).isOdd();
     }
 
     @TestEngine.Test
     public void test2() {
         System.out.println("test2(" + integerArgument + ")");
-        assertThat((Integer) integerArgument.value() % 2).isOdd();
+        assertThat(integerArgument.value() % 2).isOdd();
     }
 
     @TestEngine.AfterEach
